requirements1.txt
numpy==1.24.4
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
faiss-gpu==1.7.2
Pillow==10.2.0
tqdm==4.66.4
pandas==2.2.3
streamlit==1.45.1
altair==4.2.2
scikit-learn==1.6.1
sentencepiece==0.2.0
